requirements.txt
Django==3.1.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
django-cors-headers==3.4.0
django-filter==2.4.0
django-phonenumber-field==4.0.0
asgiref==3.3.1
celery==4.4.7
channels==3.0.2
channels-redis==3.2.0
phonenumbers==8.12.8
pillow==7.2.0
pyjwt==1.7.1
psycopg2==2.8.5
redis==3.5.3
requests==2.25.1
drf-yasg==1.17.1
django-rest-auth==0.9.5
django-tabbed-admin==1.0.4
django-debug-toolbar==3.1.1